split demo to separate files
